+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Support for varclan and varclansetting.
 - Trigger support for shiftop and onclick variants.
 - Support for "midi" type.
+- Add "Goto Script" tab to the "Find Action" window. 
 
 ### Changed